+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * X axis: time.
   * Y axis: artist playcount.
   * Color: album playcount.
-* Highlighting on selection (different color and an info box).
+* Highlighting on selection:
+  * An info box with date, artist, album and track.
+  * Different color for all scrobbles of a given artist.
+  * Date of the first scrobble of a given artist.
 * Keyboard navigation (traversing the timeline with arrow keys).
 
 ## [0.0.0] - 2019-08-23